Made the MedPacks
MedPacks can be pulled from Prefab into scene. They can be picked up by the player the same as the weapons. The player has a max amount of medpacks he can hold. This amount is adjustable in the player inspector.
The UI shows the max amount. The player prefab max amount is set to 3 ( but can be adjusted and will update UI)
The player can press the Q key to use med packs.
The medpack prefab is set to heal for 5 hp.
The player cannot use a med pack if at max hp, and cannot exceed max hp with a medpack.